Experience and reward scaling adjusted.
Completing a floor now dispenses some of the accumulated rewards(securing them so that they are not lost if the player is forced to flee later).

Dungeon exits no longer work unless the boss has been killed(as these are not yet implemented, they no longer work at all, but the code exists to support them later).